+++ src/Products/DateRangeInRangeIndex/interfaces.py
@@ -1,0 +1,4 @@
+from zope.interface import Interface
+
+class IDateRangeInRangeIndex(Interface):
+	pass